[sports] Add Sport to GameStory Model & View
Adding Sport to Game for reading. Sport should already be part of game as we
added it in the Write portion of this feature.

1/ Make sure that Game.sport property is accessible to View for GameStory
2/ Add display of this property to GameStory in Python View
3/ Add CSS to make sure it looks good. (this part needs work)

also, refactor the story generation to include a story factory that we can use
for good. forever. for.ev.er.

note that this requires the db to be cleared, since games with no sport field
in the db are fataling in SqObject._get_property() when the parameter
key='sport_id' but no corresponding field in the properties dict.

also note that the css file has a FIXME.